Add separate ebuild for PyTorch documentation
PyTorch documentation build process has its own dependencies, and
among all, it depends on PyTorch and Torchvision simultaneously,
while Torchvision depends on PyTorch itself. So the documentation
was decided to be moved to separate package. `doc` USE-flag in
PyTorch ebuild is now responsible only for Doxygen Caffe2
documentation. It's a little confusing, but seems to be the only
compliant way to have things done.