Add missing OpenCL dependency
Without the `dev-libs/clhpp` package, with `USE='opencl'`, compilation
fails with:

```
In file included from /path/to/sci-libs/pytorch-1.3.1/work/pytorch-1.3.1/caffe2/contrib/opencl/context.cc:1:
/path/to/sci-libs/pytorch-1.3.1/work/pytorch-1.3.1/caffe2/contrib/opencl/context.h:14:10: fatal error: CL/cl.hpp: No such file or directory
   14 | #include <CL/cl.hpp>
      |          ^~~~~~~~~~~
compilation terminated.
```

Note that as two ebuild files changed, the `Manifest` file needs to be
recomputed, and is *not* included in this patch to avoid conflict with
the `master` branch (The `missing-manifests` pull request should be
applied first, then this very pull request, then the `Manifest` for
`sci-libs/pytorch` should be rebuilt and committed).